requirements.txt
click==8.0.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.0
oauthlib==3.1.1
protobuf==3.19.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
Werkzeug==2.0.2
